requirements.txt
fastapi
uvicorn
pydantic
motor
beanie
python-dotenv
cryptocode
python-jose
passlib
bcrypt